# doc/api.yaml
openapi: 3.0.3
...
info:
  title: WASAPhoto
  description: Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
              upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
paths:
...
  /photos/{pid}/likes:
      parameters:
        - $ref: "#/components/parameters/photo_id"
      get:
        tags: ["Likes"]
        operationId: getLikes
        summary: Get the likes of a Photo
        description: Get the likes of a Photo
        security:
          - bearerAuth: []
        responses:
          "200":
            description: Photo Likes
            content:
              application/json:
                schema:
                  type: object
                  description: List of Likes
                  properties:
                    likes:
                      type: array
                      description: List of likes
                      minItems: 1
                      maxItems: 100
                      items:
                        $ref: "#/components/schemas/Like"
          "400" : { $ref: "#/components/responses/BadRequest" }
          "401" : { $ref: "#/components/responses/UnauthorizedError" }
          "500" : { $ref: "#/components/responses/InternalServerError" }
  /photos/{pid}/likes/{uid}:
    parameters:
      - $ref: "#/components/parameters/photo_id"
      - $ref: "#/components/parameters/user_id"
    put:
      tags: ["Likes"]
      operationId: likePhoto
      summary: Like a Photo
      description: Like a Photo
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Photo Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Like"
        "400" : { $ref: "#/components/responses/BadRequest" }
        "401" : { $ref: "#/components/responses/UnauthorizedError" }
        "500" : { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["Likes"]
      operationId: unlikePhoto
      summary: Unlike a Photo
      description: Unlike a Photo
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Photo Unliked
        "400" : { $ref: "#/components/responses/BadRequest" }
        "401" : { $ref: "#/components/responses/UnauthorizedError" }
        "500" : { $ref: "#/components/responses/InternalServerError" }
  /photos/{pid}/comments:
    parameters:
      - $ref: "#/components/parameters/photo_id"
    get:
      tags: ["Comments"]
      operationId: getComments
      summary: Get the comments of a Photo
      description: Get the comments of a Photo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Photo Comments
          content:
            application/json:
              schema:
                type: object
                description: List of Comments
                properties:
                  comments:
                    type: array
                    description: List of comments
                    minItems: 1
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/Comment"
        "400" : { $ref: "#/components/responses/BadRequest" }
        "401" : { $ref: "#/components/responses/UnauthorizedError" }
        "500" : { $ref: "#/components/responses/InternalServerError" }
  /photos/{pid}/comments/{uid}:
    parameters:
      - $ref: "#/components/parameters/photo_id"
      - $ref: "#/components/parameters/user_id"
    post:
      tags: ["Comments"]
      operationId: commentPhoto
      summary: Add a comment to a picture
      description: Add a comment to a picture
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      summary: Delete a Comment
      tags: ["Comments"]
      operationId: uncommentPhoto
      description: Delete a Comment
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Comment Deleted
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /user/{uid}/follow:
    parameters:
      - $ref: "#/components/parameters/user_id"
    get:
      operationId: getFollowedUsers
      tags: ["Follows"]
      summary: Get Followed Users
      description: Get Followed Users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Followed Users
          content:
            application/json:
              schema:
                type: object
                description: List of followed users
                properties:
                  followed_users:
                    type: array
                    description: List of followed users
                    minItems: 1
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /user/{uid}/follow/{fid}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/follow_id"
    put:
      operationId: followUser
      tags: ["Follows"]
      summary: Follow an User
      description: Follow an User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: unfollowUser
      tags: ["Follows"]
      summary: Unfollow an User
      description: Unfollow an User
      security:
        - bearerAuth: []
      responses:
        "204":
          description: User Unfollowed Successfully
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /user/{uid}/ban:
    parameters:
      - $ref: "#/components/parameters/user_id"
    get:
      operationId: getBannedUsers
      tags: ["Bans"]
      summary: Get Banned Users
      description: Get Banned Users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Banned Users
          content:
            application/json:
              schema:
                type: object
                description: List of banned users
                properties:
                  banned_users:
                    type: array
                    description: List of banned users
                    minItems: 1
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{uid}/ban/{bid}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/banned_id"  
    put:
      operationId: banUser
      tags: ["Bans"]
      summary: Ban an User
      description: Ban an User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User Banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ban"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: unbanUser
      tags: ["Bans"]
      summary: Unban an User
      description: Unban an User
      security:
        - bearerAuth: []
      responses:
        "204":
          description: User Unbanned Successfully
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  schemas:
...
    User:
      type: object
      description: User
      properties:
        user_id:
          type: string
          description: The user's identifier
        username:
          type: string
          description: The user's username
      required:
        - user_id
...
    Like:
      type: object
      description: Like
      properties:
        photo_id:
          type: string
          description: The photo's identifier
        user_id:
          type: string
          description: The user's identifier
      required:
        - photo_id
        - user_id
    Comment:
      type: object
      description: Comment
      properties:
        message:
          type: string
          description: The comment's message
        user_id:
          type: string
          description: ID of the commenting User
        photo_id:
          type: string
          description: ID of the commented Photo
      required:
        - user_id
        - photo_id
    Ban:
      type: object
      description: Ban User
      properties:
        banner_id:
          description: The user who banned
          type: string
        banned_id:
          description: The user who was banned
          type: string
      required:
        - banner_id
        - banned_id
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
    UnauthorizedError:
      description: Access token is missing or invalid
    NotFoundError:
      description: The requested resource was not found
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    photo_id:
        description: The photo's identifier
        schema:
          type: string
          pattern: '^[a-zA-Z0-9-_]+$'
          minLength: 12
          maxLength: 12
        name: pid
        in: path
        required: true
        allowEmptyValue: false
    user_id:
      description: The user's identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9-_]+$'
        minLength: 12
        maxLength: 12
      name: uid
      in: path
      required: true
      allowEmptyValue: false
    banned_id:
      description: The user's identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9-_]+$'
        minLength: 12
        maxLength: 12
      name: bid
      in: path
      required: true
      allowEmptyValue: false
    follow_id:
      description: The user's identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9-_]+$'
        minLength: 12
        maxLength: 12
      name: fid
      in: path
      required: true
      allowEmptyValue: false