# service/api/comment-api.go
package api

import (
	"encoding/json"
	"net/http"

	"github.com/fertech02/Wasa-repository/service/api/reqcontext"
	db "github.com/fertech02/Wasa-repository/service/database"
	"github.com/julienschmidt/httprouter"
)

func (rt *_router) commentPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {

	var comment db.Comment
	err := json.NewDecoder(r.Body).Decode(&comment)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	// get the photo id from the request params
	photoId := ps.ByName("pid")
	if photoId == "" {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	// get the user id from the request params
	userId := ps.ByName("uid")
	if userId == "" {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	comment.Pid = photoId
	comment.Uid = userId

	// get the message from the request body
	var message struct {
		Message string `json:"message"`
	}

	err = json.NewDecoder(r.Body).Decode(&message)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	comment.Message = message.Message

	// Add the comment in the db
	err = rt.db.Comment(&comment)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}
}

func (rt *_router) uncommentPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {

	// get the photo id from the request params
	pid := ps.ByName("pid")
	if pid == "" {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	// get the user id from the request params
	uid := ps.ByName("uid")
	if uid == "" {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	// Check if uid is authorized to delete the comment
	unauthorized, err := CheckAuthorizedId(r, uid)
	if err != nil {
		ctx.Logger.WithField("error", err).Error("Failed to check authorization")
		w.WriteHeader(http.StatusInternalServerError)
		return
	}
	if unauthorized {
		w.WriteHeader(http.StatusUnauthorized)
		return
	}

	// Delete the comment from the db
	err = rt.db.Uncomment(pid, uid)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}
}

func (rt *_router) getComments(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {

	// get the photo id from the request params
	pid := ps.ByName("pid")
	if pid == "" {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	// Get the comments from the db
	comments, err := rt.db.GetComments(pid)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	// Return the comments
	err = json.NewEncoder(w).Encode(comments)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}
}